requirements.txt
torch>=2.0.0
torchvision
torchaudio
timm
transformers
opencv-python
pillow
scikit-image
numpy
pandas
scikit-learn
matplotlib
seaborn
shap
lime
grad-cam
streamlit
reportlab
segmentation-models-pytorch
kaggle
requests
tqdm
flake8
